Docker_Full_Project/pyproject.toml
[tool.poetry]
name = ""
version = "0.1.0"
description = ""
authors = [""]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.111.0"
pydantic = "2.7.3"
uvicorn = "0.30.1"
gunicorn = "22.0.0"
pydantic-settings = "^2.3.1"
pymongo = "^4.7.3"

passlib = "1.7.4"
bcrypt = "4.0.1"
pyjwt = "2.8.0"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "7.1.2"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
